// client/src/interfaces/interface.ts
export interface ViewStateObject {
    longitude: number;
    latitude: number;
    zoom: number;
}

export interface Location{
    longitude: number;
    latitude: number;
    name: string;
    description: string;
    userID: string;
}


export interface Coordinates{
    latitude: number
    longitude: number
}